requirements.txt
torch>=1.9.0
numpy>=1.21.0
gymnasium[classic-control]>=0.28.0
matplotlib>=3.5.0
tensorboard>=2.8.0
tqdm>=4.64.0
opencv-python>=4.5.0
pygame>=2.0.0
# Isaac Sim dependencies
omni-isaac-gym>=1.0.0
omni-isaac-sim>=2023.1.0
omni-isaac-gym-envs>=1.0.0
# Additional dependencies for Isaac Sim
Pillow>=8.0.0
scipy>=1.7.0 